refactor: 도메인 로직, UI 로직 폴더 생성 및 파일 배치
