+++ src/Domain/BridgeGame.js
@@ -1,4 +1,4 @@
-const { BRIDGE } = require('./Constant');
+const { BRIDGE } = require('../Constant');
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
